requirements.txt
scikit-learn
scipy
numpy
matplotlib
pandas
statsmodels
sklearn
Pillow
tensorflow
keras